+++ example/hello/hello.dart
@@ -14,8 +14,54 @@ import 'package:dart_web_toolkit/i18n.dart' as i18n;
 import 'package:dart_web_toolkit/text.dart' as text;
 import 'package:dart_web_toolkit/scheduler.dart' as scheduler;
 
-// History managem,ent {
+// Clipped Image
 void main() {
+  ui.Label lbl = new ui.Label();
+  ui.Button btn = new ui.Button("Clip this image");
+  ui.Button btn2 = new ui.Button("Restore image");
+  
+  // Create an image, not yet referencing a URL. We make it final so that we
+  // can manipulate the image object within the ClickHandlers for the buttons.
+  ui.Image image = new ui.Image("http://www.google.com/images/logo.gif");
+
+  // Hook up an error handler, so that we can be informed if the image fails
+  // to load.
+  image.addErrorHandler(new event.ErrorHandlerAdapter((event.ErrorEvent evt) {
+    lbl.text = "An error occurred while loading.";
+  }));
+
+  // Point the image at a real URL.
+  //image.setUrl("http://www.google.com/images/logo.gif");
+
+  // When the user clicks this button, we want to clip the image.
+  btn.addClickHandler(new event.ClickHandlerAdapter((event.ClickEvent evt) {
+    image.setVisibleRect(70, 0, 47, 110);
+  }));
+  btn.setWidth("120px");
+
+  // When the user clicks this button, we want to restore the image to its
+  // unclipped state.
+  btn2.addClickHandler(new event.ClickHandlerAdapter((event.ClickEvent evt) {
+    image.setUrl("http://www.google.com/images/logo.gif");
+  }));
+  btn2.setWidth("120px");
+
+  // Add the image, label, and clip/restore buttons to the root panel.
+  ui.VerticalPanel panel = new ui.VerticalPanel();
+  panel.add(lbl);
+  panel.add(image);
+
+  ui.HorizontalPanel buttonPanel = new ui.HorizontalPanel();
+  buttonPanel.add(btn);
+  buttonPanel.add(btn2);
+
+  panel.add(buttonPanel);
+  
+  ui.RootPanel.get().add(panel);
+}
+
+// History management {
+void main_47() {
   
   ui.Label lbl = new ui.Label();